+++ Controllers/character.js
@@ -5,6 +5,9 @@ const CHARACTER = require('../models/character')
 
 exports.Create = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const filename = req.file.filename.replace(/\s+/g, '');  // Remove all spaces
 
         const hasWhitespaceInKey = obj => {
@@ -44,6 +47,9 @@ exports.Create = async function (req, res, next) {
 
 exports.Found = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const hasWhitespaceInKey = obj => {
             return Object.keys(obj).some(key => /\s/.test(key));
         };
@@ -97,6 +103,9 @@ exports.Found = async function (req, res, next) {
 
 exports.Read = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const characterData = await CHARACTER.find();
 
         res.status(200).json({
@@ -115,6 +124,9 @@ exports.Read = async function (req, res, next) {
 
 exports.Update = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const hasWhitespaceInKey = obj => {
             return Object.keys(obj).some(key => /\s/.test(key));
         };
@@ -144,6 +156,9 @@ exports.Update = async function (req, res, next) {
 
 exports.Delete = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         // Find and delete the character by ID
         const Character = await CHARACTER.findByIdAndDelete(req.params.id);
