+++ Controllers/cover.js
@@ -3,6 +3,9 @@ const USER = require('../models/users')
 
 exports.Create = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const filename = req.file.filename.replace(/\s+/g, '');  // Remove all spaces
         // req.body.CoverURL = `https://lolcards.link/api/public/images/cover/${filename}`;
 
@@ -43,6 +46,9 @@ exports.Create = async function (req, res, next) {
 
 exports.Emoji = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const hasWhitespaceInKey = obj => {
             return Object.keys(obj).some(key => /\s/.test(key));
         };
@@ -99,6 +105,9 @@ exports.Emoji = async function (req, res, next) {
 
 exports.Realistic = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const hasWhitespaceInKey = obj => {
             return Object.keys(obj).some(key => /\s/.test(key));
         };
@@ -148,6 +157,9 @@ exports.Realistic = async function (req, res, next) {
 
 exports.Read = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const coverData = await COVER.find();
 
         res.status(200).json({
@@ -166,6 +178,9 @@ exports.Read = async function (req, res, next) {
 
 exports.Update = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         const hasWhitespaceInKey = obj => {
             return Object.keys(obj).some(key => /\s/.test(key));
         };
@@ -195,6 +210,9 @@ exports.Update = async function (req, res, next) {
 
 exports.Delete = async function (req, res, next) {
     try {
+        if (req.headers['x-forwarded-proto'] !== 'https') {
+            throw new Error('Please use HTTPS protocol')
+        }
         await COVER.findByIdAndDelete(req.params.id);
         res.status(204).json({
             status: 1,
